refactor: _update 파일 병합.
