--- src/data/dtos/trip/tripScheduleResponseDto_update.ts
@@ -1,10 +1,0 @@
-// Server -> Client
-export interface TripScheduleResponseDto {
-  id: number;
-  title: string;
-  destination: string;
-  start_date: string;
-  end_date: string;
-  members: string[];
-  created_by: string;
-}
