+++ src/data/dtos/user/userDto.ts
@@ -2,8 +2,8 @@ export interface UserDto {
   id: string;
   provider: string;
   email: string;
-  user_image: string;
+  user_image?: string;
   nickname: string;
-  user_memo: string;
-  trip_history: number[];
+  user_memo?: string;
+  created_at: Date;
 }
