+++ src/data/integrations/passport/kakaoOAuth.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import { Strategy as KakaoStrategy } from 'passport-kakao';
-import { UserRepository } from '../../../domain/repositories/userRepository_update';
+import { UserRepository } from '../../../domain/repositories/userRepository';
 
 const app = express();
