+++ src/domain/models/user.ts
@@ -2,10 +2,10 @@ export interface User {
   id: string;
   provider: string;
   email: string;
-  user_image: string;
+  user_image?: string;
   nickname: string;
-  user_memo: string;
+  user_memo?: string;
   access_token: string;
   refresh_token: string;
-  trip_history: number[];
+  created_at: Date;
 }
