--- src/domain/models/user_update.ts
@@ -1,11 +1,0 @@
-export interface User {
-  id: string;
-  provider: string;
-  email: string;
-  user_image?: string;
-  nickname: string;
-  user_memo?: string;
-  access_token: string;
-  refresh_token: string;
-  created_at: Date;
-}
