+++ src/domain/repositories/userRepository.ts
@@ -1,13 +1,16 @@
-import { User } from '../models/user_update';
+import { User } from '../models/user'; // _update 파일은 db 붙이기 작성중인 파일
+
 export interface UserRepository {
   createUser: (userData: Omit<User, 'id'>) => Promise<User>;
   findUserById: (id: string) => Promise<User | undefined>;
   findUserByEmail: (email: string) => Promise<User | undefined>;
   findUsersByEmail: (email: string) => Promise<User[] | undefined>;
+  findUsersByEmails: (email: string[]) => Promise<User[]>;
   findUserByEmailAndProvider: (
     email: string,
     provider: string,
   ) => Promise<User | undefined>;
+  findUsersByEmailContains: (query: string) => Promise<User[]>;
   updateTokens: (
     email: string,
     access_token: string,
@@ -25,7 +28,15 @@ export interface UserRepository {
   updateUserMemo: (id: string, user_memo: string) => Promise<User | undefined>;
 
   deleteUser: (id: string) => Promise<boolean>;
-  updateUserTripHistory: (userId: string, tripId: number) => Promise<boolean>; // 함수 표현식 방식
-  // updateUserTripHistory2(userId: string, tripId: number): Promise<void>; // 메서드 방식
-  removeTripFromHistory: (userId: string, tripId: number) => Promise<boolean>;
+
+  // ✅ 중간 테이블을 고려한 추가 기능 (User가 참여한 TripSchedule 관리)
+  getUserTripSchedules: (user_id: string) => Promise<number[]>; // User가 속한 TripSchedule ID 조회
+  addUserToTripSchedule: (
+    user_id: string,
+    tripSchedule_id: number,
+  ) => Promise<void>; // User를 TripSchedule에 추가
+  removeUserFromTripSchedule: (
+    user_id: string,
+    tripSchedule_id: number,
+  ) => Promise<void>; // TripSchedule에서 User 제거
 }
