+++ src/index.ts
@@ -7,11 +7,11 @@ import path from 'path';
 
 import initPassport from './data/integrations/passport/initPassport';
 import { rootRouter } from './presentation/routers';
-import { di_update } from './di_update';
+import { di } from './di';
 
 const app = express();
 
-di_update(app);
+di(app);
 
 //적용 env 체크
 const env = process.env.NODE_ENV || 'local';
