+++ src/presentation/routers/auth/googleRouter.ts
@@ -1,8 +1,7 @@
 import dotenv from 'dotenv';
 import { Router } from 'express';
 import passport from 'passport';
-// import { User } from '../../../domain/models/user';
-import { User } from '../../../domain/models/user_update';
+import { User } from '../../../domain/models/user';
 
 const env = process.env.NODE_ENV || 'local';
